Use double input timing; tweak ship turn speed
Convert input timing values to double precision for more accurate duration tracking: change getLastPressDuration return type and press/release time variables in InputManager, and update KeyState to use double timestamps. Reduce ship turnSpeed from 5.0f to 3.0f for gameplay tuning. Update logger/log entry headers and implementation and refresh game.log with a new run output.

Files changed: src/input/InputManager.hpp/cpp, src/input/KeyState.hpp, src/ecs/systems/ShipControlSystem.hpp, src/utilities/logger/LogEntry.hpp, src/utilities/logger/Logger.cpp/hpp, game.log.

+++ src/input/InputManager.cpp
@@ -9,14 +9,14 @@
 
 #include "../utilities/logger/Logger.hpp"
 
-float InputManager::getLastPressDuration(SDL_Scancode key) const {
+double InputManager::getLastPressDuration(SDL_Scancode key) const {
 	// make sure in bounds
-	if (key >= SDL_SCANCODE_COUNT) return 0.0f;
+	if (key >= SDL_SCANCODE_COUNT) return 0.0;
 	// make sure the key has been pressed at least once
-	if (keyboardState_[key].totalPressCount == 0) return 0.0f;
+	if (keyboardState_[key].totalPressCount == 0) return 0.0;
 	// if the key is currently down, return the current time
 	if (keyboardState_[key].isDown) {
-		return (static_cast<float>(SDL_GetTicks()) / 1000.0f) - keyboardState_[key].lastPressTime;
+		return (SDL_GetTicks() / 1000.0) - keyboardState_[key].lastPressTime;
 	}
 	// finally just give the data
 	return keyboardState_[key].lastReleaseTime - keyboardState_[key].lastPressTime;
@@ -28,7 +28,7 @@ void InputManager::processEvent(const SDL_Event& event) {
 		SDL_Scancode key = event.key.scancode;
 		
 		if (!keyboardState_[key].isDown) {
-			float pressTime = (static_cast<float>(SDL_GetTicks())) / 1000.0f;
+			double pressTime = SDL_GetTicks() / 1000.0;
 
 			// KEY_DOWN block, after updating state:
 			LOG_DEBUG("Input", "KEY DOWN: {}", static_cast<int>(key));
@@ -51,7 +51,7 @@ void InputManager::processEvent(const SDL_Event& event) {
 	else if (event.type == SDL_EVENT_KEY_UP) {
 		// convienice variable
 		SDL_Scancode key = event.key.scancode;
-		float pressTime = (static_cast<float>(SDL_GetTicks())) / 1000.0f;
+		double pressTime = SDL_GetTicks() / 1000.0;
 
 		// KEY_UP block, after updating state:
 		LOG_DEBUG("Input", "KEY UP: {}", static_cast<int>(key));
